refactor: split cmd_inspect, clean/validate, schema_ops; fix csv_read; add toolkit init, inspect url --run
- fix(csv_read): error messages now interpolate actual values
- fix(csv_read): add iso-8859-1 and ascii encoding aliases
- fix(csv_read): validate nullstr list entries must be strings
- feat(cli): add top-level 'toolkit init' command (run init alias)
- feat(cli): 'inspect url --run' for single-step URL probe + scaffold + init
- refactor(cmd_inspect): split 537 LOC into toolkit/cli/inspect/ package
  (paths_ops, schema_diff_ops, url_ops, probe_ops, _helpers)
- refactor(clean/validate): split 490 LOC into _column_rules.py, _helpers.py
- refactor(mcp/schema_ops): extract helpers into toolkit/mcp/_schema_utils.py
- refactor(cli): remove dead code cmd_scout_url.py re-export
- tests: add 212-line test_csv_read.py for csv_read bugfixes
