+++ toolkit/cli/app.py
@@ -11,6 +11,7 @@ from toolkit.cli.cmd_inspect import register as register_inspect
 from toolkit.cli.cmd_scaffold import register as register_scaffold
 from toolkit.cli.cmd_batch import register as register_batch
 from toolkit.cli.cmd_blocker_hints import register as register_blocker_hints
+from toolkit.cli.cmd_init import register as register_init
 
 app = typer.Typer(no_args_is_help=True, add_completion=False)
 
@@ -24,6 +25,7 @@ register_inspect(app)
 register_scaffold(app)
 register_batch(app)
 register_blocker_hints(app)
+register_init(app)
 
 
 def main():
